refactor(post): extract the component
